pyproject.toml
[tool.poetry]
name = "DevRewind"
version = "0.1.0"
description = ""
authors = ["williamfzc <williamfzc@foxmail.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "dev_rewind" }
]

[tool.poetry.dependencies]
python = "^3.8.1,!=3.9.7"
pydantic = "^2.4.2"
gitpython = "^3.1.38"
loguru = "^0.7.2"
langchain = "^0.0.315"
click = "^8.1.7"
chromadb = "^0.4.14"
tiktoken = "^0.5.1"
streamlit = "^1.27.2"
rapidfuzz = "^3.4.0"
openai = "^0.28.1"
tinydb = "^4.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
devr = 'dev_rewind.cli.click:main'
devr_web = 'dev_rewind.cli.streamlit.entry:main'
